% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnModelSelection.R
\name{AICFunction}
\alias{AICFunction}
\title{Model Section Via Akaike Information Criterion}
\usage{
AICFunction(
  logLikelihood,
  nParameters,
  clusterRunOutput = NA,
  gmin,
  gmax,
  parallel = FALSE
)
}
\arguments{
\item{logLikelihood}{A vector with value of final log-likelihoods for
each cluster size.}

\item{nParameters}{A vector with number of parameters for each
cluster size.}

\item{clusterRunOutput}{Output from mplnVariational, mplnMCMCParallel, or
mplnMCMCNonParallel, if available. Default value is NA. If provided,
the vector of cluster labels obtained by mclust::map() for best model
will be provided in the output.}

\item{gmin}{A positive integer specifying the minimum number of components
to be considered in the clustering run.}

\item{gmax}{A positive integer, >gmin, specifying the maximum number of
components to be considered in the clustering run.}

\item{parallel}{TRUE or FALSE indicating if MPLNClust::mplnMCMCParallel
has been used.}
}
\value{
Returns an S3 object of class MPLN with results.
\itemize{
\item allAICvalues - A vector of AIC values for each cluster size.
\item AICmodelselected - An integer specifying model selected by AIC.
\item AICmodelSelectedLabels - A vector of integers specifying cluster labels
for the model selected. Only provided if user input clusterRunOutput.
\item AICMessage - A character vector indicating if spurious clusters are
detected. Otherwise, NA.
}
}
\description{
Performs model selection using Akaike Information Criterion (AIC).
Formula: - 2 * logLikelihood + 2 * nParameters.
}
\examples{
# Generating simulated data
# Not run
# trueMu1 <- c(6.5, 6, 6, 6, 6, 6)
# trueMu2 <- c(2, 2.5, 2, 2, 2, 2)
# trueSigma1 <- diag(6) * 2
# trueSigma2 <- diag(6)
# sampleData <- mplnDataGenerator(nObservations = 100,
#                                 dimensionality = 6,
#                                 mixingProportions = c(0.79, 0.21),
#                                 mu = rbind(trueMu1, trueMu2),
#                                 sigma = rbind(trueSigma1, trueSigma2),
#                                 produceImage = "No")
# Clustering
# mplnResults <- mplnMCMCParallel(dataset = sampleData$dataset,
#                                 membership = sampleData$trueMembership,
#                                 gmin = 1,
#                                 gmax = 2,
#                                 nChains = 3,
#                                 nIterations = 1000,
#                                 initMethod = "kmeans",
#                                 nInitIterations = 2,
#                                 normalize = "Yes")
#
# Model selection
# AICmodel <- AICFunction(logLikelihood = mplnResults$logLikelihood,
#                         nParameters = mplnResults$numbParameters,
#                         clusterRunOutput = mplnResults$allResults,
#                         gmin = mplnResults$gmin,
#                         gmax = mplnResults$gmax,
#                         parallel = TRUE)

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.
}
\author{
{Anjali Silva, \email{anjali.silva@uhnresearch.ca}}
}
