% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnMCMCEMClustering.R
\name{BICFunction}
\alias{BICFunction}
\title{Model Selection Via Bayesian Information Criterion}
\usage{
BICFunction(
  logLikelihood,
  nParameters,
  nObservations,
  clusterRunOutput = NA,
  gmin,
  gmax,
  parallel = FALSE
)
}
\arguments{
\item{logLikelihood}{A vector with value of final log-likelihoods for
each cluster size.}

\item{nParameters}{A vector with number of parameters for each
cluster size.}

\item{nObservations}{A positive integer specifying the number of observations
in the dataset analyzed.}

\item{clusterRunOutput}{Output from mplnVariational, mplnMCMCParallel, or
mplnMCMCNonParallel, if available. Default value is NA. If provided,
the vector of cluster labels obtained by mclust::map() for best model
will be provided in the output.}

\item{gmin}{A positive integer specifying the minimum number of components
to be considered in the clustering run.}

\item{gmax}{A positive integer, >gmin, specifying the maximum number of
components to be considered in the clustering run.}

\item{parallel}{TRUE or FALSE indicating if MPLNClust::mplnMCMCParallel
has been used.}
}
\value{
Returns an S3 object of class MPLN with results.
\itemize{
\item allBICvalues - A vector of BIC values for each cluster size.
\item BICmodelselected - An integer specifying model selected by BIC
\item BICmodelSelectedLabels - A vector of integers specifying cluster labels
for the model selected. Only provided if user input clusterRunOutput.
\item BICMessage - A character vector indicating if spurious clusters are
detected. Otherwise, NA.
}
}
\description{
Performs model selection using Bayesian Information Criterion (BIC) by
Schwarz (1978). Formula: - 2 * logLikelihood + (nParameters * log(nObservations)).
}
\examples{
trueMu1 <- c(6.5, 6, 6, 6, 6, 6)
trueMu2 <- c(2, 2.5, 2, 2, 2, 2)

trueSigma1 <- diag(6) * 2
trueSigma2 <- diag(6)

# Generating simulated data
sampleData <- MPLNClust::mplnDataGenerator(nObservations = 100,
                                 dimensionality = 6,
                                 mixingProportions = c(0.79, 0.21),
                                 mu = rbind(trueMu1, trueMu2),
                                 sigma = rbind(trueSigma1, trueSigma2),
                                 produceImage = "No")

# Clustering
mplnResults <- MPLNClust::mplnVariational(dataset = sampleData$dataset,
                                membership = sampleData$trueMembership,
                                gmin = 1,
                                gmax = 2,
                                initMethod = "kmeans",
                                nInitIterations = 2,
                                normalize = "Yes")

# Model selection
BICmodel <- MPLNClust::BICFunction(logLikelihood = mplnResults$logLikelihood,
                         nParameters = mplnResults$numbParameters,
                         nObservations = nrow(mplnResults$dataset),
                         clusterRunOutput = mplnResults$allResults,
                         gmin = mplnResults$gmin,
                         gmax = mplnResults$gmax,
                         parallel = FALSE)

}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.
}
\author{
{Anjali Silva, \email{anjali.silva@uhnresearch.ca}}
}
