% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpln.R
\name{mpln}
\alias{mpln}
\title{Clustering Via Mixtures of Multivariate Poisson-Log Normal Distributions}
\usage{
mpln(
  dataset,
  membership,
  gmin = 1,
  gmax = 2,
  nChains = 3,
  nIterations = NA,
  initMethod = "kmeans",
  nInitIterations = 0,
  normalize = "Yes"
)
}
\arguments{
\item{dataset}{A dataset of class matrix and type integer such that
rows correspond to observations and columns correspond to variables.}

\item{membership}{A numeric vector of length nrow(dataset) containing the
cluster membership of each observation.}

\item{gmin}{A positive integer indicating the minimum number of components
to be considered in the clustering run.}

\item{gmax}{A positive integer, >gmin, indicating the maximum number of
components to be considered in the clustering run.}

\item{nChains}{A positive integer specifying the number of Markov chains.
The default is 3, the recommended minimum number.}

\item{nIterations}{A negative value of class "numeric" indicating}

\item{initMethod}{An algorithm for initialization. Current options are
"kmeans", "random", "medoids", "clara", or "fanny".}

\item{nInitIterations}{A negative value of class "numeric" indicating}

\item{normalize}{A negative value of class "numeric" indicating}
}
\description{
Performs clustering using mixtures of multivariate Poisson-log
normal distribution and model selection using AIC, AIC3, BIC and ICL.
}
