% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnVisualize.R
\name{mplnVisualize}
\alias{mplnVisualize}
\title{Visualize Clustered Results Via MPLN}
\usage{
mplnVisualize(
  dataset,
  clusterMembershipVector,
  name = "",
  plots = "all",
  Xpanels = NA,
  Ypanels = NA,
  format = "pdf"
)
}
\arguments{
\item{dataset}{A dataset of class matrix and type integer such that
rows correspond to observations and columns correspond to variables.}

\item{clusterMembershipVector}{A numeric vector of length nrow(dataset)
containing the cluster membership of each observation as generated by
mpln().}

\item{name}{Name for the plot being generated.}

\item{plots}{A character string indicating which plots to be produced.
Options are 'heatmaps', 'lines', and 'all'. Default is 'all'.}

\item{Xpanels}{A positive number indicating the number of panels on x axis.
Default NA.}

\item{Ypanels}{A positive number indicating the number of panels on y axis.
Default NA.}

\item{format}{Character string indicating the format of the image to
be produced. Default 'pdf'.}
}
\value{
Plotting function provides the possibility for line and heatmap plots.
}
\description{
A function to visualize the clustering results obtained from a
mixtures of multivariate Poisson-log normal (MPLN) model.
}
