% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnMCMCEMClustering.R
\name{mplnDataGenerator}
\alias{mplnDataGenerator}
\title{Generating Data Using Mixtures of MPLN}
\usage{
mplnDataGenerator(
  nObservations,
  dimensionality,
  mixingProportions,
  mu,
  sigma,
  produceImage = "No",
  ImageName = "sampleName"
)
}
\arguments{
\item{nObservations}{A positive integer indicating the number of observations for
the dataset.}

\item{dimensionality}{A positive integer indicating the dimensionality for the
dataset.}

\item{mixingProportions}{A numeric vector that length equal to the number of total
components, indicating the proportion of each component. Vector content should
sum to 1.}

\item{mu}{A matrix of size (dimensionality x number of components), indicating
the mean for each component. See example.}

\item{sigma}{A matrix of size ((dimensionality * number of components) x
dimensionality), indicating the covariance matrix for each component.
See example.}

\item{produceImage}{A character string indicating whether or not to
produce an image. Options "Yes" or "No". Image will be produced as
'Pairs plot of log-transformed data.png" in the current working
directory.}

\item{ImageName}{A character string indicating name for image, if
produceImage is set to "Yes". Default is "TwoComponents".}
}
\value{
Returns an S3 object of class mplnDataGenerator with results.
\itemize{
\item dataset - Simulated dataset.
\item trueMembership -A numeric vector indicating the membership of
each observation.
\item probaPost - A matrix indicating the posterior probability that
each observation belong to the component/cluster.
\item truenormfactors - A numeric vector indicating the true
normalization factors used for adjusting the library sizes.
\item observations - Number of observations in the simulated dataset.
\item dimensionality - Dimensionality of the simulated dataset.
\item mixingProportions - A numeric vector indicating the mixing
proportion of each component.
\item mu - True mean used for the simulated dataset.
\item sigma - True covariances used for the simulated dataset.
}
}
\description{
This function simulates data from a mixture of MPLN model.
}
\examples{
trueMu1 <- c(6.5, 6, 6, 6, 6, 6)
trueMu2 <- c(2, 2.5, 2, 2, 2, 2)

trueSigma1 <- diag(6) * 2
trueSigma2 <- diag(6)

# Generating simulated data
sampleData <- MPLNClust::mplnDataGenerator(nObservations = 100,
                                           dimensionality = 6,
                                           mixingProportions = c(0.79, 0.21),
                                           mu = rbind(trueMu1, trueMu2),
                                           sigma = rbind(trueSigma1, trueSigma2),
                                           produceImage = "No",
                                           ImageName = "TwoComponents")

}
\references{
Aitchison, J. and C. H. Ho (1989). The multivariate Poisson-log normal distribution.
\emph{Biometrika} 76.

Silva, A. et al. (2019). A multivariate Poisson-log normal mixture model
for clustering transcriptome sequencing data. \emph{BMC Bioinformatics} 20.
\href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-2916-0}{Link}
}
\author{
Anjali Silva, \email{a.silva@utoronto.ca}
}
