% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnVisualize.R
\name{mplnVisualize}
\alias{mplnVisualize}
\title{Visualize Clustered Results Via MPLN}
\usage{
mplnVisualize(
  dataset,
  clusterMembershipVector = NA,
  fileName = "FileName",
  plots = "all",
  format = "pdf"
)
}
\arguments{
\item{dataset}{A dataset of class matrix and type integer such that
rows correspond to observations and columns correspond to variables.}

\item{clusterMembershipVector}{A numeric vector of length nrow(dataset)
containing the cluster membership of each observation as generated by
mpln(). Default NA.}

\item{fileName}{Name for the plot being generated.}

\item{plots}{A character string indicating which plots to be produced.
Options are 'heatmaps', 'lines', and 'all'. Default is 'all'.}

\item{format}{Character string indicating the format of the image to
be produced. Default 'pdf'. Options 'pdf' or 'png'.}
}
\value{
Plotting function provides the possibility for line and heatmap
plots.
}
\description{
A function to visualize the clustering results obtained from a
mixtures of multivariate Poisson-log normal (MPLN) model. This
function produces heatmaps and line plots.
}
\examples{
# Generating simulated data
# Not run
# trueMu1 <- c(6.5, 6, 6, 6, 6, 6)
# trueMu2 <- c(2, 2.5, 2, 2, 2, 2)

# trueSigma1 <- diag(6) * 2
# trueSigma2 <- diag(6)

# simulatedCounts <- mplnDataGenerator(nObservations = 70,
#                                       dimensionality = 6,
#                                       mixingProportions = c(0.79, 0.21),
#                                       mu = rbind(trueMu1, trueMu2),
#                                       sigma = rbind(trueSigma1, trueSigma2),
#                                       produceImage = "No")

# MPLNClustResults <- mpln(dataset = simulatedCounts$dataset,
#                          membership = NA,
#                          gmin = 2,
#                          gmax = 2,
#                          nChains = 3,
#                          nIterations = 100,
#                          initMethod = "kmeans",
#                          nInitIterations = 0,
#                          normalize = "Yes")

# MPLNVisuals <- mplnVisualize(dataset = simulatedCounts$dataset,
#                              clusterMembershipVector =
#                  MPLNClustResults$all_results[[1]]$all_results$clusterlabels,
#                               fileName = 'TwoClusterModel', plots = 'all',
#                               format = 'png')

}
\author{
Anjali Silva, \email{anjali.silva@uhnresearch.ca}
}
