% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnVisualize.R
\name{mplnVisualizeAlluvial}
\alias{mplnVisualizeAlluvial}
\title{Alluvial Plot of Multiple Clustering Results}
\usage{
mplnVisualizeAlluvial(
  nObservations = 50L,
  firstGrouping = floor(runif(50, min = 1, max = 8)),
  secondGrouping = vector(mode = "integer", length = 0),
  thirdGrouping = vector(mode = "integer", length = 0),
  fourthGrouping = vector(mode = "integer", length = 0),
  fileName = paste0("Plot_", date()),
  printPlot = TRUE,
  format = "pdf"
)
}
\arguments{
\item{nObservations}{An integer specifying the total number of
observations, N, in the dataset. Default value is 50L.}

\item{firstGrouping}{A vector of length nObservations, specifying the
cluster membership of observations. This must be provided. Colors
will be assigned based on cluster membership provided in this
vector. Default value is a vector of length 50.}

\item{secondGrouping}{A vector of length N, specifying the cluster
membership of N observations. This could be obtained via another
clustering run or from a different model selection criteria.
Default value is an empty vector.}

\item{thirdGrouping}{A vector of length N, specifying the cluster
membership of N observations. This could be obtained via another
clustering run or from a different model selection criteria.
Default value is an empty vector.}

\item{fourthGrouping}{A vector of length N, specifying the cluster
membership of N observations. This could be obtained via another
clustering run or from a different model selection criteria.
Default value is an empty vector.}

\item{fileName}{Unique character string indicating the name for the plot
being generated. Default is Plot_date, where date is obtained from
date().}

\item{printPlot}{Logical indicating if plot(s) should be saved in local
directory. Default TRUE. Options TRUE or FALSE.}

\item{format}{Character string indicating the format of the image to
be produced. Default 'pdf'. Options 'pdf' or 'png'.}
}
\value{
An alluvial plot should be returned. The x-axis values
are in the order of vectors assigned (if any) to firstGrouping,
secondGrouping, thirdGrouping and fourthGrouping, respectively.
}
\description{
A function to visualize clustering results via alluvial plots,
using the alluvial::alluvial() function. The function produces
an alluvial plot provided multiple clustering results for
the same group of observations. Up to four varying results
could be visualized. Minimum, one clustering result for
visualization is required. Maximum 10 colors (clusters) are
supported. Colors are assigned based on cluster membership
assigned for argument 'firstGrouping'.
}
\examples{
# Example 1
# Assign values for parameters
trueMu1 <- c(6.5, 6, 6, 6, 6, 6)
trueMu2 <- c(2, 2.5, 2, 2, 2, 2)

trueSigma1 <- diag(6) * 2
trueSigma2 <- diag(6)

# Generate simulated data for 500 x 6 dataset
simulatedCounts <- MPLNClust::mplnDataGenerator(nObservations = 500,
                                      dimensionality = 6,
                                      mixingProportions = c(0.79, 0.21),
                                      mu = rbind(trueMu1, trueMu2),
                                      sigma = rbind(trueSigma1, trueSigma2),
                                      produceImage = "No")

 # Clustering data for G = 1:2
 MPLNClustResults <- MPLNClust::mplnVariational(
                              dataset = as.matrix(simulatedCounts$dataset),
                              membership = "none",
                              gmin = 1,
                              gmax = 2,
                              initMethod = "kmeans",
                              nInitIterations = 1,
                              normalize = "Yes")

 # Visualize clustering results using alluvial plot
 # Access results using models selected via model selection criteria
 alluvialPlot <- mplnVisualizeAlluvial(nObservations = nrow(simulatedCounts$dataset),
                           firstGrouping = MPLNClustResults$BICresults$BICmodelSelectedLabels,
                           secondGrouping = MPLNClustResults$ICLresults$ICLmodelSelectedLabels,
                           thirdGrouping = MPLNClustResults$AIC3results$AIC3modelSelectedLabels,
                           fourthGrouping = MPLNClustResults$AICresults$AICmodelSelectedLabels,
                           fileName = paste0('Plot_',date()),
                           printPlot = FALSE,
                           format = 'pdf')

 # Example 2
 # Perform clustering via K-means with centers = 2
 # Visualize clustering results using alluvial plot for
 # K-means and above MPLNClust results. Note, coloring
 # is set with respect to argument firstGrouping, which is
 # assinged MPLNClust results.

 set.seed(1234)
 alluvialPlotMPLNClust <- mplnVisualizeAlluvial(nObservations = nrow(simulatedCounts$dataset),
                               firstGrouping = MPLNClustResults$BICresults$BICmodelSelectedLabels,
                               secondGrouping = kmeans(simulatedCounts$dataset, 2)$cluster,
                               fileName = paste0('Plot_',date()),
                               printPlot = FALSE,
                               format = 'pdf')

 # Note, coloring is set with respect to argument firstGrouping,
 # which is assinged K-means results.
 set.seed(1234)
 alluvialPlotKmeans <- mplnVisualizeAlluvial(nObservations = nrow(simulatedCounts$dataset),
                           firstGrouping = kmeans(simulatedCounts$dataset, 2)$cluster,
                           secondGrouping = MPLNClustResults$BICresults$BICmodelSelectedLabels,
                           fileName = paste0('Plot_',date()),
                           printPlot = FALSE,
                           format = 'pdf')

}
\references{
Bojanowski,  M., R. Edwards (2016). alluvial: R Package for
Creating Alluvial Diagrams. R package version 0.1-2.
\href{https://github.com/mbojan/alluvial}{Link}

Wickham, H., R. François, L. Henry and K. Müller (2021).
dplyr: A Grammar of Data Manipulation. R package version
1.0.7. \href{https://CRAN.R-project.org/package=dplyr}{Link}
}
\author{
Anjali Silva, \email{a.silva@utoronto.ca}
}
