% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnVisualize.R
\name{mplnVisualizeBar}
\alias{mplnVisualizeBar}
\title{Visualize Posterior Probabilities via Bar Plot}
\usage{
mplnVisualizeBar(
  dataset,
  probabilities = NA,
  clusterMembershipVector = NA,
  fileName = paste0("Plot_", date()),
  LinePlotColours = "black",
  printPlot = TRUE,
  format = "pdf"
)
}
\arguments{
\item{dataset}{A dataset of class matrix and type integer such that
rows correspond to observations (N) and columns (C) correspond to variables.}

\item{probabilities}{A matrix of size N x C, such that rows correspond
to N observations and columns correspond to C clusters. Each row
should sum to 1. Default is NA.}

\item{clusterMembershipVector}{A numeric vector of length nrow(dataset)
containing the cluster membership of each observation. Default is NA.}

\item{fileName}{Unique character string indicating the name for the plot
being generated. Default is Plot_date, where date is obtained from
date().}

\item{printPlot}{Logical indicating if plot(s) should be saved in local
directory. Default TRUE. Options TRUE or FALSE.}

\item{format}{Character string indicating the format of the image to
be produced. Default 'pdf'. Options 'pdf' or 'png'.}
}
\value{
A bar plot of posterior probabilities for each observation
in the dataset.
}
\description{
A function to produce a barplot of posterior probabilities
for each observation, after clustering via mixtures of
multivariate Poisson-log normal (MPLN) model.
}
\examples{
# Example 1
trueMu1 <- c(6.5, 6, 6, 6, 6, 6)
trueMu2 <- c(2, 2.5, 2, 2, 2, 2)

trueSigma1 <- diag(6) * 2
trueSigma2 <- diag(6)

# Generating simulated data
simulatedCounts <- MPLNClust::mplnDataGenerator(nObservations = 100,
                                      dimensionality = 6,
                                      mixingProportions = c(0.79, 0.21),
                                      mu = rbind(trueMu1, trueMu2),
                                      sigma = rbind(trueSigma1, trueSigma2),
                                      produceImage = "No")

 # Clustering data
 MPLNClustResults <- MPLNClust::mplnVariational(
                              dataset = as.matrix(simulatedCounts$dataset),
                              membership = "none",
                              gmin = 1,
                              gmax = 2,
                              initMethod = "kmeans",
                              nInitIterations = 1,
                              normalize = "Yes")

 # Visualize posterior probabilities via a bar plot
 MPLNVisuals <- MPLNClust::mplnVisualizeBar(dataset = simulatedCounts$dataset,
                                            probabilities =
                                            MPLNClustResults$allResults$`G=2`$probaPost,
                                            clusterMembershipVector =
                                            MPLNClustResults$allResults$`G=2`$clusterlabels,
                                            fileName = 'PosteriorProbMPLN',
                                            printPlot = FALSE,
                                            format = 'png')

}
\author{
Anjali Silva, \email{a.silva@utoronto.ca}
}
