% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMPLNClust.R
\name{runMPLNClust}
\alias{runMPLNClust}
\title{Launch the shiny app for package MPLNClust}
\usage{
runMPLNClust()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runMPLNClust()
}

}
